Code improved
